refactor: extract main tests to sibling file
